Fix the character ignoring issue
First the for loop didn't use every character in the characters variable to generate the password. It only used the first few characters in that variable according to the number in the input field.

Ex: If we enter 10 as the password length in the input field, the password was generated only by using the first 10 characters of the characters variable.

This led the for loop to ignore the other characters in the characters variable.

First I was using the value input field (password length) to multiply the math random value.

I fixed this issue by using the length of the characters variable to multiply the math random.